No two sdk packages
