+++ Makefile
@@ -29,6 +29,8 @@ rm-mocked:
 rm-builders:
 	# not sure why, but go.tmpl is matched when run in makefile but not when run in the commandline.
 	grep -rl "^// Code generated by github.com/smartcontractkit/chainlink-common/pkg/capabilities/cli" | grep .go$ | grep -v go.tmpl$ | xargs -r rm
+	# test wasm file should be removed as well
+	rm -f ./pkg/workflows/wasm/host/test/cmd/testmodule.wasm
 
 .PHONY: generate
 generate: mockery install-protoc gomods
