--- pkg/workflows/wasm/generate_wasm.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-GOOS=wasip1 GOARCH=wasm go build -o ./test/cmd/testmodule.wasm ./test/cmd/main.go
