+++ pkg/workflows/wasm/host/test/generate_wasm.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+GOOS=wasip1 GOARCH=wasm go build -o ./test/cmd/testmodule.wasm ./test/cmd/main.go
