+++ pkg/workflows/wasm/sdk.go
@@ -1,0 +1,7 @@
+package wasm
+
+import "github.com/smartcontractkit/chainlink-common/pkg/workflows/sdk"
+
+type Runtime struct{}
+
+var _ sdk.Runtime = (*Runtime)(nil)
