--- pkg/workflows/wasm/sdk/sdk.go
@@ -1,7 +1,0 @@
-package sdk
-
-import "github.com/smartcontractkit/chainlink-common/pkg/workflows/sdk"
-
-type Runtime struct{}
-
-var _ sdk.Runtime = (*Runtime)(nil)
